Log a helpful message
The error message "Some top level structure elements in the current ruleset are missing functional metadata of type 'recordIdentifier' which is required for OPAC import!" can be hard to debug. This patch adds logging of additional helpful messages. The control flow is changed so that in case of error, all error cases are found and logged, not just the first one.