Fix HibernateException: could not re-associate uninitialized transient
... collection". This is caused because the object being saved has a
lazy-loading collection that is not loaded, but it is no longer attached
to a session. The solution is to merge (re-attach) the object to the
current session before doing something with it again.

The merge operation cannot modify the current object, but returns a new
(proxy) object that must be used from now on. To make this
counterintuitive behaviour of the API more obvious, the CheckReturnValue
annotation was introduced, which is evaluated by Spotbugs, which results
in a Spotbugs alert if it is ignored.